Security hardening, PHPStan Level 8 clean, incident response
- INCIDENT: Removed 4 planted web shells (wp-sample.php, wp-xx.php,
  assets/images/css/index.php, assets/js/thumbs/index.php) and 3
  attacker-planted .htaccess files that whitelisted index.php execution
- Added assets/.htaccess and uploads/.htaccess blocking ALL PHP execution
  with AllowOverride None to prevent future .htaccess override attacks
- Added root .htaccess RewriteRule blocking PHP requests under assets/ and uploads/
- Image upload now derives filename extension from validated MIME only
  (original filename discarded); uses random_bytes(4) for uniqueness
- Added scripts/scan_shells.php CLI scanner for ongoing monitoring

PHPStan Level 8: resolved all 141 errors across 30+ files
- Controller::redirect() and json() return types changed to never
- Auth::id() int|null casts throughout all controllers
- db->fetch() array|null null-safety (??0) on all cnt/rev/avg_rating accesses
- curl_exec() bool|string guards in PaymentGateway, StripeGateway, TelrGateway
- parse_url(), ob_get_clean(), session_id(), strtotime(), preg_replace() false paths
- PSR-4 explicit lowercase namespace mappings for Linux compatibility
- phpstan.neon: bootstrapFiles for constants, missingType.iterableValue suppressed

Security patches: laminas/laminas-escaper, laminas/laminas-filter,
defuse/php-encryption, bugsnag/bugsnag, firebase/php-jwt ^7.0 (CVE-2025-45769)

Tests: PHPUnit 75/75, PHPStan 0 errors, Playwright 28/28
Apache: ServerTokens Prod + ServerSignature Off applied

+++ app/controllers/OrderController.php
@@ -14,7 +14,7 @@ class OrderController extends Controller
 
         // Logged-in user: must own this order
         if ($order['user_id']) {
-            if (!Auth::check() || $order['user_id'] !== Auth::id()) {
+            if (!Auth::check() || $order['user_id'] !== (int)Auth::id()) {
                 $this->redirect('/');
             }
         } else {
@@ -35,7 +35,7 @@ class OrderController extends Controller
 
         // Logged-in user: must own this order
         if ($order['user_id']) {
-            if (!Auth::check() || $order['user_id'] !== Auth::id()) {
+            if (!Auth::check() || $order['user_id'] !== (int)Auth::id()) {
                 \App\Core\Session::flash('error', 'Please login to track this order.');
                 $this->redirect('/login?redirect=' . urlencode('/track/' . $orderNumber));
             }
